+++ src/lib.rs
@@ -6,6 +6,8 @@
 //! core implementation. To use a pre-configured stable vector, use
 //! [`StableVec`].
 //!
+//! This crate uses `#![no_std]` but requires the `alloc` crate.
+//!
 //!
 //! # Why?
 //!
@@ -95,11 +97,42 @@
 #![deny(missing_debug_implementations)]
 #![deny(intra_doc_link_resolution_failure)]
 
+// ----- Deal with `no_std` stuff --------------------------------------------
+// Ideally we could use the crate `no-std-compat` for this. Unfortunately, it
+// is only usable on nightly at the moment. Once that changes, all of this will
+// be replaced.
 
-use std::{
+#![no_std]
+
+// Import the real `std` for tests.
+#[cfg(test)]
+#[macro_use]
+extern crate std;
+
+// Otherwise import `core` and `alloc` and define a module `std` which
+// simulates the real `std` by reexporting the symbols from `core` and `alloc`.
+// Well, not perfectly, but all parts we need.
+#[cfg(not(test))] extern crate alloc as _alloc;
+#[cfg(not(test))] extern crate core as _core;
+#[cfg(not(test))]
+mod std {
+    pub use _core::*;
+    pub use _alloc::alloc;
+
+    pub mod prelude {
+        pub mod v1 {
+            pub use _core::prelude::v1::*;
+            pub use _alloc::vec::Vec;
+        }
+    }
+}
+// ---------------------------------------------------------------------------
+
+
+use crate::std::{
+    prelude::v1::*,
     cmp,
     fmt,
-    io,
     iter::FromIterator,
     mem,
     ops::{Index, IndexMut},
@@ -1580,22 +1613,6 @@ impl<T, C: Core<T>> Extend<T> for StableVecFacade<T, C> {
             self.push(elem);
         }
     }
-}
-
-/// Write into `StableVecFacade<u8>` by appending `u8` elements. This is
-/// equivalent to calling `push` for each byte.
-impl<C: Core<u8>> io::Write for StableVecFacade<u8, C> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.extend_from_slice(buf);
-        Ok(buf.len())
-    }
-
-    fn write_all(&mut self, buf: &[u8]) -> io::Result<()> {
-        self.extend_from_slice(buf);
-        Ok(())
-    }
-
-    fn flush(&mut self) -> io::Result<()> { Ok(()) }
 }
 
 impl<'a, T, C: Core<T>> IntoIterator for &'a StableVecFacade<T, C> {