+++ src/tests.rs
@@ -1,4 +1,5 @@
 use std::{
+    prelude::v1::*,
     fmt::Debug,
     panic::RefUnwindSafe,
 };
@@ -140,7 +141,10 @@ macro_rules! assert_sv_eq {
 
 macro_rules! gen_tests_for {
     ($ty:ident) => {
-        use std::iter::FromIterator;
+        use std::{
+            prelude::v1::*,
+            iter::FromIterator,
+        };
         use quickcheck_macros::quickcheck;
         use super::assert_sv_eq_fn;
 
@@ -888,23 +892,6 @@ macro_rules! gen_tests_for {
             assert_eq!(sv.remove_last(), Some('d'));
             sv.extend_from_slice(&['e']);
             assert_sv_eq!(sv, [0 => 'a', 1 => 'b', 2 => 'c', 4 => 'e']);
-        }
-
-        #[test]
-        fn write() {
-            use std::io::Write;
-
-            let mut sv = $ty::new();
-
-            sv.write_all(&[0, 7, 3]).unwrap();
-            assert_sv_eq!(sv, [0 => 0, 1 => 7, 2 => 3]);
-
-            sv.remove_last();
-            sv.write_all(&[4, 8]).unwrap();
-            assert_sv_eq!(sv, [0 => 0, 1 => 7, 3 => 4, 4 => 8]);
-
-            sv.write_all(&[5]).unwrap();
-            assert_sv_eq!(sv, [0 => 0, 1 => 7, 3 => 4, 4 => 8, 5 => 5]);
         }
 
         #[test]